sort of fix how builtins work.
some builtins (cd , exit) need to run in the shell process in order to have
any effect.  others (help, ls, ...) can run just fine in a child process.
add flags to builtins and check for them in both places.

this is still not fantastic, but I need to do some research and figure out
a better way to handle this.